+++ CHANGELOG.md
@@ -10,6 +10,7 @@ description: |-
 #### v0.2.0 (unreleased)
 - Provide support for Public DNS service
 - Support cloud monitoring for DB instances/clusters
+- Add security_group argument to DB instance/cluster resources
 
 #### v0.1.16
 - Add config option to run against clouds with old cloud containers API